Add logging UserService and Unit testing UserService

+++ src/main/java/by/tms/swager/service/UserService.java
@@ -7,6 +7,7 @@ import by.tms.swager.exception.UserNotFoundException;
 import by.tms.swager.exception.InvalidTokenException;
 import by.tms.swager.repository.TokenRepository;
 import by.tms.swager.repository.UserRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,6 +16,7 @@ import java.util.UUID;
 
 @Service
 @Transactional
+@Slf4j
 public class UserService {
     private final UserRepository userRepository;
     private final TokenRepository tokenRepository;
@@ -28,51 +30,91 @@ public class UserService {
         Optional<User> userByUsername = userRepository.findUserByUsername(user.getUsername());
 
         if(userByUsername.isPresent()) {
+            log.warn("Create account. Login is busy | login: " + user.getUsername());
             throw new LoginIsBusyException();
         }
 
         userRepository.save(user);
+        log.info("Create account. Success | User: " + user);
     }
 
-    public void userCreate(User[] users) {
+    public void userCreate(User[] users) throws LoginIsBusyException {
         for (User user : users) {
-            userRepository.save(user);
+            userCreate(user);
         }
     }
 
     public void userDelete(String username) throws UserNotFoundException {
+        if(!userRepository.existsUserByUsername(username)) {
+            log.warn("Delete user. Invalid username | username: " + username);
+            throw new UserNotFoundException();
+        }
         userRepository.deleteUserByUsername(username);
+        log.info("Delete user. Success | login: " + username);
     }
 
     public void userUpdate(String username, User user) throws UserNotFoundException {
-        long findId = userRepository.findUserByUsername(username).orElseThrow(UserNotFoundException::new).getId();
-        user.setId(findId);
+        Optional<User> userByUsername = userRepository.findUserByUsername(username);
+        if(userByUsername.isEmpty()) {
+            log.warn("Update user. Invalid username | username: " + username);
+            throw new UserNotFoundException();
+        }
+
+        user.setId(userByUsername.get().getId());
+        user.setUsername(userByUsername.get().getUsername());
         userRepository.save(user);
+        log.info("Update user. Success | User: " + user);
     }
 
     public User getUser(String username) throws UserNotFoundException {
-        return userRepository.findUserByUsername(username).orElseThrow(UserNotFoundException::new);
+        Optional<User> userByUsername = userRepository.findUserByUsername(username);
+        if (userByUsername.isPresent()) {
+            log.info("Get user. Success | User: " + userByUsername.get());
+            return userByUsername.get();
+        }
+        log.warn("Get user. Invalid username | username: " + username);
+        throw new UserNotFoundException();
     }
 
     public String loginUser(String username, String password) throws UserNotFoundException {
-        User findUser = userRepository.findUserByUsernameAndPassword(username, password).orElseThrow(UserNotFoundException::new);
-        Optional<Token> findToken = tokenRepository.findTokenByUserId(findUser.getId());
+        System.out.println(username);
+        System.out.println(password);
+        Optional<User> userByUsernameAndPassword = userRepository.findUserByUsernameAndPassword(username, password);
+        if(userByUsernameAndPassword.isEmpty()) {
+            log.warn("Login user. invalid username or password | username: " + username + ", password: " + password);
+            throw new UserNotFoundException();
+        }
 
+        Optional<Token> findToken = tokenRepository.findTokenByUserId(userByUsernameAndPassword.get().getId());
         if(findToken.isPresent()) {
-            return findToken.get().getId().toString();
+            String token = findToken.get().getId().toString();
+            log.info("Login user. Success | user: " + userByUsernameAndPassword.get() + ". Get token: " + token);
+            return token;
         } else {
-            Token newToken = new Token(findUser);
-            newToken = tokenRepository.save(newToken);
+            Token newToken = tokenRepository.save(new Token(userByUsernameAndPassword.get()));
+            log.info("Login user. Success | user: " + userByUsernameAndPassword.get() + ". Create new token: " + newToken.getId().toString());
             return newToken.getId().toString();
         }
     }
 
-    public void logoutUser(UUID token) throws UserNotFoundException, InvalidTokenException {
-        checkAuthorizationUser(token);
+    public void logoutUser(UUID token) throws InvalidTokenException {
+        if (!tokenRepository.existsTokenById(token)) {
+            log.warn("Logout user. Invalid token | token: " + token);
+            throw new InvalidTokenException();
+        }
+
         tokenRepository.deleteById(token);
+        log.info("Logout user. Success | token: " + token);
     }
 
     public User checkAuthorizationUser(UUID token) throws UserNotFoundException {
-        return tokenRepository.findById(token).orElseThrow(UserNotFoundException::new).getUser();
+        Optional<Token> findToken = tokenRepository.findById(token);
+        if (findToken.isPresent()) {
+            log.debug("Check authorization. Success | user: " + findToken.get().getUser() + ", token: " + token);
+            return findToken.get().getUser();
+        }
+
+        log.warn("Check authorization. Invalid token | token: " + token);
+        throw new UserNotFoundException();
     }
 }
